Move Array extension into Array.swift
